Rename devMode to ResetService
