+++ misk/src/main/kotlin/misk/reset/ResetCoordinationService.kt
@@ -1,4 +1,4 @@
-package misk.devmode
+package misk.reset
 
 import com.google.common.util.concurrent.AbstractIdleService
 import com.google.common.util.concurrent.Service
@@ -6,29 +6,28 @@ import com.google.inject.Inject
 import com.google.inject.Key
 import com.google.inject.Singleton
 import misk.DependentService
-import misk.inject.toKey
 
 @Singleton
-class DevModeCoordinationService @Inject internal constructor(
-  @DevMode val isDevMode: Boolean,
-  @DevModeService val devModeServices: List<Service>
+class ResetCoordinationService @Inject internal constructor(
+  @RunResetService val runResetService: Boolean,
+  @ResetService val resetServices: List<Service>
 ) : AbstractIdleService(), DependentService {
 
-  val devModeServiceKeys = devModeServices.map { Key.get(it::class.java).typeLiteral }
+  val devModeServiceKeys = resetServices.map { Key.get(it::class.java).typeLiteral }
 
   override val consumedKeys: Set<Key<*>> =
-      devModeServices.fold(setOf()) { consumedKeysSet, service ->
+      resetServices.fold(setOf()) { consumedKeysSet, service ->
         consumedKeysSet + (service as DependentService).consumedKeys.filter {
           // TODO(nb): don't filter out keys with matching type but different annotations
           !devModeServiceKeys.contains(it.typeLiteral)
         }
       }
 
-  override val producedKeys: Set<Key<*>> = setOf(Key.get(DevModeCoordinationService::class.java))
+  override val producedKeys: Set<Key<*>> = setOf(Key.get(ResetCoordinationService::class.java))
 
   override fun startUp() {
-    if (isDevMode) {
-      devModeServices.forEach {
+    if (runResetService) {
+      resetServices.forEach {
         it.startAsync()
         it.awaitRunning()
       }
@@ -36,8 +35,8 @@ class DevModeCoordinationService @Inject internal constructor(
   }
 
   override fun shutDown() {
-    if (isDevMode) {
-      devModeServices.forEach { it.stopAsync() }
+    if (runResetService) {
+      resetServices.forEach { it.stopAsync() }
     }
   }
 }
