+++ misk/src/main/kotlin/misk/reset/ResetService.kt
@@ -1,7 +1,7 @@
-package misk.devmode
+package misk.reset
 
 import javax.inject.Qualifier
 
 @Qualifier
 @Target(AnnotationTarget.FIELD, AnnotationTarget.FUNCTION, AnnotationTarget.VALUE_PARAMETER)
-annotation class DevModeService
+annotation class ResetService
